Added more commands
Added acquire2CMD, DropCMD, TurnCMD, and ReloadCMD. modifed drop command in LifeForm so a dropped weapon would be added to cell the life form was in, or would fail to drop if no space available. also added reload command to human which reloads the current weapon they are holding if any. began attack command, started adding method to Environment to locate nearest LifeForm to the player in the direction they are facing.