Finished findNearest
Finished findNearest() method in Environment and tests. will find closest life form to passed life form in a cone infront of the focus base on their direction. cone starts at 1 cell infront and is 3 wide and grows 1 cell in both left and right for each cell forward that is checked. so a cell facing west at (5,0) will check in order {(5,1), (4,1), (6,1); (5,2), (4,2), (6,2), (3,2), (7,2)....} . for 2 life forms the same distance away, the lifeForm to the left of the focus wil be returned